refactored chart code into separate Chart component
